+++ zetaclient/evm_client.go
@@ -64,29 +64,29 @@ const (
 // Filled with above constants depending on chain
 type EVMChainClient struct {
 	*ChainMetrics
-	chain                     common.Chain
-	evmClient                 EVMRPCClient
-	KlaytnClient              KlaytnRPCClient
-	zetaClient                ZetaCoreBridger
-	Tss                       TSSSigner
-	lastBlockScanned          uint64
-	lastBlock                 uint64
-	BlockTimeExternalChain    uint64 // block time in seconds
-	txWatchList               map[ethcommon.Hash]string
-	Mu                        *sync.Mutex
-	db                        *gorm.DB
-	outTxPendingTransaction   map[string]*ethtypes.Transaction
-	outTXConfirmedReceipts    map[string]*ethtypes.Receipt
-	outTXConfirmedTransaction map[string]*ethtypes.Transaction
-	MinNonce                  int64
-	MaxNonce                  int64
-	OutTxChan                 chan OutTx // send to this channel if you want something back!
-	stop                      chan struct{}
-	fileLogger                *zerolog.Logger // for critical info
-	logger                    EVMLog
-	cfg                       *config.Config
-	params                    observertypes.CoreParams
-	ts                        *TelemetryServer
+	chain                      common.Chain
+	evmClient                  EVMRPCClient
+	KlaytnClient               KlaytnRPCClient
+	zetaClient                 ZetaCoreBridger
+	Tss                        TSSSigner
+	lastBlockScanned           uint64
+	lastBlock                  uint64
+	BlockTimeExternalChain     uint64 // block time in seconds
+	txWatchList                map[ethcommon.Hash]string
+	Mu                         *sync.Mutex
+	db                         *gorm.DB
+	outTxPendingTransactions   map[string]*ethtypes.Transaction
+	outTXConfirmedReceipts     map[string]*ethtypes.Receipt
+	outTXConfirmedTransactions map[string]*ethtypes.Transaction
+	MinNonce                   int64
+	MaxNonce                   int64
+	OutTxChan                  chan OutTx // send to this channel if you want something back!
+	stop                       chan struct{}
+	fileLogger                 *zerolog.Logger // for critical info
+	logger                     EVMLog
+	cfg                        *config.Config
+	params                     observertypes.CoreParams
+	ts                         *TelemetryServer
 
 	BlockCache *lru.Cache
 }
@@ -123,9 +123,9 @@ func NewEVMChainClient(
 	ob.zetaClient = bridge
 	ob.txWatchList = make(map[ethcommon.Hash]string)
 	ob.Tss = tss
-	ob.outTxPendingTransaction = make(map[string]*ethtypes.Transaction)
+	ob.outTxPendingTransactions = make(map[string]*ethtypes.Transaction)
 	ob.outTXConfirmedReceipts = make(map[string]*ethtypes.Receipt)
-	ob.outTXConfirmedTransaction = make(map[string]*ethtypes.Transaction)
+	ob.outTXConfirmedTransactions = make(map[string]*ethtypes.Transaction)
 	ob.OutTxChan = make(chan OutTx, 100)
 
 	logFile, err := os.OpenFile(ob.chain.ChainName.String()+"_debug.log", os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0600)
@@ -296,7 +296,7 @@ func (ob *EVMChainClient) Stop() {
 // returns: isIncluded, isConfirmed, Error
 // If isConfirmed, it also post to ZetaCore
 func (ob *EVMChainClient) IsSendOutTxProcessed(sendHash string, nonce uint64, cointype common.CoinType, logger zerolog.Logger) (bool, bool, error) {
-	if !ob.hasTxConfirmed(nonce) {
+	if !ob.isTxConfirmed(nonce) {
 		return false, false, nil
 	}
 	params := ob.GetCoreParams()
@@ -558,11 +558,7 @@ func (ob *EVMChainClient) observeOutTx() {
 	if err != nil || timeoutNonce <= 0 {
 		timeoutNonce = 100 * 3 // process up to 100 hashes
 	}
-	rpcRestTime, err := strconv.Atoi(os.Getenv("OS_RPC_REST_TIME"))
-	if err != nil || rpcRestTime <= 0 {
-		rpcRestTime = 20 // 20ms
-	}
-	ob.logger.ObserveOutTx.Info().Msgf("observeOutTx using timeoutNonce %d seconds, rpcRestTime %d ms", timeoutNonce, rpcRestTime)
+	ob.logger.ObserveOutTx.Info().Msgf("observeOutTx using timeoutNonce %d seconds", timeoutNonce)
 
 	ticker, err := NewDynamicTicker(fmt.Sprintf("EVM_observeOutTx_%d", ob.chain.ChainId), ob.GetCoreParams().OutTxTicker)
 	if err != nil {
@@ -587,7 +583,7 @@ func (ob *EVMChainClient) observeOutTx() {
 				if nonceInt < lowestOutTxNonceToObserve[ob.chain.ChainId] {
 					continue
 				}
-				if ob.hasTxConfirmed(nonceInt) { // Go to next tracker if this one already has a confirmed tx
+				if ob.isTxConfirmed(nonceInt) { // Go to next tracker if this one already has a confirmed tx
 					continue
 				}
 				for _, txHash := range tracker.HashList {
@@ -596,13 +592,11 @@ func (ob *EVMChainClient) observeOutTx() {
 						ob.logger.ObserveOutTx.Warn().Msgf("observeOutTx timeout on chain %d nonce %d", ob.chain.ChainId, nonceInt)
 						break TRACKERLOOP
 					default:
-						err := ob.confirmTxByHash(txHash.TxHash, nonceInt)
-						time.Sleep(time.Duration(rpcRestTime) * time.Millisecond)
-						if err == nil { // confirmed
+						if ob.confirmTxByHash(txHash.TxHash, nonceInt) {
 							ob.logger.ObserveOutTx.Info().Msgf("observeOutTx confirmed outTx %s for chain %d nonce %d", txHash.TxHash, ob.chain.ChainId, nonceInt)
 							break
 						}
-						ob.logger.ObserveOutTx.Debug().Err(err).Msgf("error confirmTxByHash: chain %s hash %s", ob.chain.String(), txHash.TxHash)
+						ob.logger.ObserveOutTx.Debug().Msgf("observeOutTx outTx %s for chain %d nonce %d not confirmed yet", txHash.TxHash, ob.chain.ChainId, nonceInt)
 					}
 				}
 			}
@@ -617,14 +611,14 @@ func (ob *EVMChainClient) observeOutTx() {
 // SetPendingTx sets the pending transaction in memory
 func (ob *EVMChainClient) SetPendingTx(nonce uint64, transaction *ethtypes.Transaction) {
 	ob.Mu.Lock()
-	ob.outTxPendingTransaction[ob.GetTxID(nonce)] = transaction
+	ob.outTxPendingTransactions[ob.GetTxID(nonce)] = transaction
 	ob.Mu.Unlock()
 }
 
 // GetPendingTx gets the pending transaction from memory
 func (ob *EVMChainClient) GetPendingTx(nonce uint64) *ethtypes.Transaction {
 	ob.Mu.Lock()
-	transaction := ob.outTxPendingTransaction[ob.GetTxID(nonce)]
+	transaction := ob.outTxPendingTransactions[ob.GetTxID(nonce)]
 	ob.Mu.Unlock()
 	return transaction
 }
@@ -632,9 +626,9 @@ func (ob *EVMChainClient) GetPendingTx(nonce uint64) *ethtypes.Transaction {
 // SetTxNReceipt sets the receipt and transaction in memory
 func (ob *EVMChainClient) SetTxNReceipt(nonce uint64, receipt *ethtypes.Receipt, transaction *ethtypes.Transaction) {
 	ob.Mu.Lock()
-	delete(ob.outTxPendingTransaction, ob.GetTxID(nonce)) // remove pending transaction, if any
+	delete(ob.outTxPendingTransactions, ob.GetTxID(nonce)) // remove pending transaction, if any
 	ob.outTXConfirmedReceipts[ob.GetTxID(nonce)] = receipt
-	ob.outTXConfirmedTransaction[ob.GetTxID(nonce)] = transaction
+	ob.outTXConfirmedTransactions[ob.GetTxID(nonce)] = transaction
 	ob.Mu.Unlock()
 }
 
@@ -642,37 +636,38 @@ func (ob *EVMChainClient) SetTxNReceipt(nonce uint64, receipt *ethtypes.Receipt,
 func (ob *EVMChainClient) GetTxNReceipt(nonce uint64) (*ethtypes.Receipt, *ethtypes.Transaction) {
 	ob.Mu.Lock()
 	receipt := ob.outTXConfirmedReceipts[ob.GetTxID(nonce)]
-	transaction := ob.outTXConfirmedTransaction[ob.GetTxID(nonce)]
+	transaction := ob.outTXConfirmedTransactions[ob.GetTxID(nonce)]
 	ob.Mu.Unlock()
 	return receipt, transaction
 }
 
-// hasTxConfirmed returns true if there is a confirmed tx for 'nonce'
-func (ob *EVMChainClient) hasTxConfirmed(nonce uint64) bool {
+// isTxConfirmed returns true if there is a confirmed tx for 'nonce'
+func (ob *EVMChainClient) isTxConfirmed(nonce uint64) bool {
 	ob.Mu.Lock()
-	confirmed := ob.outTXConfirmedReceipts[ob.GetTxID(nonce)] != nil && ob.outTXConfirmedTransaction[ob.GetTxID(nonce)] != nil
+	confirmed := ob.outTXConfirmedReceipts[ob.GetTxID(nonce)] != nil && ob.outTXConfirmedTransactions[ob.GetTxID(nonce)] != nil
 	ob.Mu.Unlock()
 	return confirmed
 }
 
 // confirmTxByHash checks if a txHash is confirmed and saves transaction and receipt in memory
-// returns nil if confirmed or error otherwise
-func (ob *EVMChainClient) confirmTxByHash(txHash string, nonce uint64) error {
+// returns true if confirmed or false otherwise
+func (ob *EVMChainClient) confirmTxByHash(txHash string, nonce uint64) bool {
 	ctxt, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
 	// query transaction
 	transaction, isPending, err := ob.evmClient.TransactionByHash(ctxt, ethcommon.HexToHash(txHash))
 	if err != nil {
-		return err
+		log.Error().Err(err).Msgf("confirmTxByHash: TransactionByHash error, txHash %s nonce %d", txHash, nonce)
+		return false
 	}
 	if transaction == nil { // should not happen
 		log.Error().Msgf("confirmTxByHash: transaction is nil for txHash %s nonce %d", txHash, nonce)
-		return fmt.Errorf("transaction is nil")
+		return false
 	}
 	if isPending { // save pending transaction
 		ob.SetPendingTx(nonce, transaction)
-		return fmt.Errorf("confirmTxByHash: txHash %s nonce %d is pending", txHash, nonce)
+		return false
 	}
 
 	// query receipt
@@ -681,32 +676,33 @@ func (ob *EVMChainClient) confirmTxByHash(txHash string, nonce uint64) error {
 		if err != ethereum.NotFound {
 			log.Warn().Err(err).Msgf("confirmTxByHash: TransactionReceipt error, txHash %s nonce %d", txHash, nonce)
 		}
-		return err
+		return false
 	}
 	if receipt == nil { // should not happen
 		log.Error().Msgf("confirmTxByHash: receipt is nil for txHash %s nonce %d", txHash, nonce)
-		return fmt.Errorf("receipt is nil")
+		return false
 	}
 
 	// check nonce and confirmations
 	if transaction.Nonce() != nonce {
 		log.Error().Msgf("confirmTxByHash: txHash %s nonce mismatch: wanted %d, got tx nonce %d", txHash, nonce, transaction.Nonce())
-		return fmt.Errorf("outtx nonce mismatch")
+		return false
 	}
 	confHeight := receipt.BlockNumber.Uint64() + ob.GetCoreParams().ConfirmationCount
 	if confHeight >= math.MaxInt64 {
-		return fmt.Errorf("confirmTxByHash: confHeight is out of range")
+		log.Error().Msgf("confirmTxByHash: confHeight is too large for txHash %s nonce %d", txHash, nonce)
+		return false
 	}
 	if confHeight > ob.GetLastBlockHeight() {
 		log.Info().Msgf("confirmTxByHash: txHash %s nonce %d included but not confirmed: receipt block %d, current block %d",
 			txHash, nonce, receipt.BlockNumber, ob.GetLastBlockHeight())
-		return fmt.Errorf("included but not confirmed")
+		return false
 	}
 
 	// confirmed, save receipt and transaction
 	ob.SetTxNReceipt(nonce, receipt, transaction)
 
-	return nil
+	return true
 }
 
 // SetLastBlockHeightScanned set last block height scanned (not necessarily caught up with external block; could be slow/paused)
@@ -1249,7 +1245,7 @@ func (ob *EVMChainClient) BuildTransactionsMap() error {
 		if err != nil {
 			return err
 		}
-		ob.outTXConfirmedTransaction[transaction.Identifier] = trans
+		ob.outTXConfirmedTransactions[transaction.Identifier] = trans
 	}
 	return nil
 }